+++ src/Resources/contao/languages/de/tl_content.php
@@ -1,21 +1,27 @@
 <?php 
 
-$GLOBALS['TL_LANG']['tl_content']['adresse_legend'] = 'Einstellungen';
-$GLOBALS['TL_LANG']['tl_content']['adresse_id'] = array('Adresse auswählen', 'Bitte wählen Sie die anzuzeigende Adresse aus.');
-$GLOBALS['TL_LANG']['tl_content']['adresse_funktion'] = array('Funktion/Amt', 'Erscheint vor dem Namen in Fettschrift.');
-$GLOBALS['TL_LANG']['tl_content']['adresse_zusatz'] = array('Zusatztext', 'Erscheint zwischen Funktion/Amt und Name. HTML ist erlaubt.');
-$GLOBALS['TL_LANG']['tl_content']['adresse_selectmails'] = array('Nur bestimmte E-Mail-Adressen anzeigen', 'Anzeige der E-Mail-Adressen einschränken');
-$GLOBALS['TL_LANG']['tl_content']['adresse_mails'] = array('E-Mail-Adressen auswählen', 'Nur aktivierte E-Mail-Adressen werden angezeigt.');
+$GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_main_legend'] = 'Schnittstelle, Saison und Liga';
+$GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_api'] = array('Schritt 1: Schnittstelle wählen', 'Bitte wählen Sie eine Schnittstelle aus. Die API-Schlüssel für die Schnittstellen müssen Sie in den Systemeinstellungen eintragen.');
+$GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_saison'] = array('Schritt 2: Saison wählen', 'Bitte wählen Sie eine Saison aus.');
+$GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_liga'] = array('Schritt 3: Liga wählen', 'Bitte wählen Sie eine Liga aus.');
 
-$GLOBALS['TL_LANG']['tl_content']['adressefoto_legend'] = 'Foto';
-$GLOBALS['TL_LANG']['tl_content']['adresse_addImage'] = array('Foto anzeigen', 'Foto im Frontend anzeigen');
-$GLOBALS['TL_LANG']['tl_content']['adresse_bildvorschau'] = array('Zugeordnetes Foto', 'Das Bild wird in der Frontendgröße angezeigt.');
-$GLOBALS['TL_LANG']['tl_content']['adresse_singleSRC'] = array('Foto ersetzen durch:', 'Datei aus dem Dateisystem wählen');
+$GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_options_legend'] = 'Anzeigeart und Mannschaft';
+$GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_typ'] = array('Anzeige', 'Bitte wählen Sie die Anzeige aus.');
+$GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_mannschaft'] = array('Mannschaft (optional)', 'Bitte wählen Sie eine Mannschaft aus (optional).');
 
-$GLOBALS['TL_LANG']['tl_content']['adresstemplate_legend'] = 'Alternatives Template';
-$GLOBALS['TL_LANG']['tl_content']['adresse_alttemplate'] = array('Aktivieren', 'Alternatives Template aktivieren');
-$GLOBALS['TL_LANG']['tl_content']['adresse_tpl'] = array('Template', 'Bitte wählen Sie das zu verwendende Frontend-Template aus.');
+// Mögliche APIs festlegen
+$GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_apis'] = array();
+if($GLOBALS['TL_CONFIG']['ergebnisdienst_apikey_bl']) $GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_apis']['1'] = 'Bundesliga';
+if($GLOBALS['TL_CONFIG']['ergebnisdienst_apikey_dsol']) $GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_apis']['2'] = 'Deutsche Schach-Online-Liga';
 
-// Sonstige 
-$GLOBALS['TL_LANG']['tl_content']['adresse_bildvorschau_leer'] = 'Fehler: Pfad nicht gefunden!';
-$GLOBALS['TL_LANG']['tl_content']['adresse_nichtaktiv'] = ' (deaktiviert)';
+// Anzeigearten festlegen
+$GLOBALS['TL_LANG']['tl_content']['ergebnisdienst_anzeigearten'] = array
+(
+	'1'  => 'Mannschaften (ohne Aufstellung)',
+	'11' => 'Mannschaften (mit Aufstellung an Brett 1-4)',
+	'12' => 'Mannschaften (mit Aufstellung an Brett 1-8)',
+	'13' => 'Mannschaften (mit kompletter Aufstellung)',
+	'2'  => 'Termine',
+	'3'  => 'Ergebnisse',
+	'4'  => 'Tabelle'
+);