+++ src/Resources/contao/languages/de/tl_settings.php
@@ -14,10 +14,10 @@
 /**
  * legends
  */
-$GLOBALS['TL_LANG']['tl_settings']['adressen_legend'] = 'Adressen';
+$GLOBALS['TL_LANG']['tl_settings']['ergebnisdienst_legend'] = 'DSB-Ergebnisdienst';
 
 /**
  * fields
  */
-$GLOBALS['TL_LANG']['tl_settings']['adressen_ImageSize'] = array('Bildgröße', 'Größe des Standardbildes');
-$GLOBALS['TL_LANG']['tl_settings']['adressen_defaultImage'] = array('Standardbild', 'Standardbild für Adressen');
+$GLOBALS['TL_LANG']['tl_settings']['ergebnisdienst_apikey_bl'] = array('API-Schlüssel Bundesliga', 'Schlüssel der JSON-Schnittstelle des Bundesliga-Ergebnisdienstes');
+$GLOBALS['TL_LANG']['tl_settings']['ergebnisdienst_apikey_dsol'] = array('API-Schlüssel DSOL', 'Schlüssel der JSON-Schnittstelle des DSOL-Ergebnisdienstes');