+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # DSB-Ergebnisdienst Changelog
 
+## Version 0.0.2 (2021-01-11)
+
+* Add: Übersetzungen tl_settings und tl_content
+* Add: Anpassung der Klasse Ergebnisdienst
+* Add: API-Feld für DSOL
+
 ## Version 0.0.1 (2020-12-07)
 
 * Initiale Version